+++ tests/LimitStreamTest.php
@@ -1,8 +1,10 @@
 <?php
 namespace GuzzleHttp\Tests\Http;
 
+use GuzzleHttp\Stream\FnStream;
 use GuzzleHttp\Stream\Stream;
 use GuzzleHttp\Stream\LimitStream;
+use GuzzleHttp\Stream\NoSeekStream;
 
 /**
  * @covers GuzzleHttp\Stream\LimitStream
@@ -74,6 +76,19 @@ class LimitStreamTest extends \PHPUnit_Framework_TestCase
         $this->assertNotSame($data, $newData);
     }
 
+    /**
+     * @expectedException \GuzzleHttp\Stream\Exception\SeekException
+     * @expectedExceptionMessage Could not seek the stream to position 2
+     */
+    public function testThrowsWhenCurrentGreaterThanOffsetSeek()
+    {
+        $a = Stream::factory('foo_bar');
+        $b = new NoSeekStream($a);
+        $c = new LimitStream($b);
+        $a->getContents();
+        $c->setOffset(2);
+    }
+
     public function testClaimsConsumedWhenReadLimitIsReached()
     {
         $this->assertFalse($this->body->eof());
@@ -90,5 +105,22 @@ class LimitStreamTest extends \PHPUnit_Framework_TestCase
     {
         $body = new LimitStream(Stream::factory('foobazbar'), 3, 3);
         $this->assertEquals('baz', $body->getContents());
+    }
+
+    public function testReturnsNullIfSizeCannotBeDetermined()
+    {
+        $a = new FnStream([
+            'getSize' => function () { return null; },
+            'tell'    => function () { return 0; },
+        ]);
+        $b = new LimitStream($a);
+        $this->assertNull($b->getSize());
+    }
+
+    public function testLengthLessOffsetWhenNoLimitSize()
+    {
+        $a = Stream::factory('foo_bar');
+        $b = new LimitStream($a, -1, 4);
+        $this->assertEquals(3, $b->getSize());
     }
 }